requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.7
Django==4.2.6
factory-boy==3.3.0
Faker==19.12.0
filelock==3.12.4
gunicorn==21.2.0
identify==2.5.30
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.5.0
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.5
